+++ src/pages/form/upload/_FormUploadPage.tsx
@@ -1,22 +1,19 @@
-import { ReactNode, useEffect, useReducer, useState } from 'react';
-import { Box } from '@components/Box';
 import 'react-datepicker/dist/react-datepicker.css';
+import { useEffect, useReducer, useState } from 'react';
+import { Box } from '@components/Box';
 import { Period } from '@src/types/period-picker/period';
-import { TbPlus } from 'react-icons/tb';
-import { FormUpload, Question, QuestionType } from '@api/form-upload/types';
+import { FormUpload, Question } from '@api/form-upload/types';
 import { SubmitButton } from '@components/SubmitButton';
-import { RiListRadio } from 'react-icons/ri';
-import { RxCross2 } from 'react-icons/rx';
-import { BsTextareaT } from 'react-icons/bs';
-import {
-  QuestionTypeButton,
-  QuestionTypeButtonChangeEvent,
-} from './QuestionTypeButton';
-import { Header } from './Header';
 import { useTopBarStore } from '@stores/topBar-stores';
-import { PiTrashLight } from 'react-icons/pi';
 import { useModal } from '@hooks/useModal';
 import { useCreateForm } from '@query-hooks/form-upload';
+import { FormUploadActionType } from '@/types/form';
+import { FormHeader } from './FormHeader';
+import { QuestionHeader } from './QuestionHeader';
+import { QuestionInput } from './components/input/QuestionInput';
+import { AddQuestionButton } from './components/button/AddQuestionButton';
+import { SubjectiveInput } from './components/input/SubjectiveInput';
+import { OptionInputList } from './components/input/OptionInputList';
 
 const initialQuestionState: Question = {
   tag: 'SUBJECTIVE',
@@ -24,21 +21,52 @@ const initialQuestionState: Question = {
   description: '',
   options: [''],
 };
-
-type ActionType = {
-  type: string;
-  questionIndex?: number;
-  questionType?: QuestionType;
-  inputName?: string;
-  inputValue?: string;
-  optionIndex?: number;
-  optionValue?: string;
+const initialFormUploadInfo = {
+  title: '',
+  description: '',
+  postId: 0,
+  startTime: '',
+  endTime: '',
+  surveyItems: [],
 };
 
 export const FormUploadPage = () => {
   const [questionList, dispatch] = useReducer(reducer, [initialQuestionState]);
+  const formUploadInfoState = useState<FormUpload>(initialFormUploadInfo);
+  const [formUploadInfo, setFormUploadInfo] = formUploadInfoState;
+  const periodState = useState<Period>({
+    start: new Date(),
+    end: new Date(),
+  });
+  const { mutate: createForm } = useCreateForm();
+  const { open } = useModal();
 
-  function reducer(state: Question[], action: ActionType): Question[] {
+  const uploadForm = () => {
+    setFormUploadInfo((prev) => ({ ...prev, surveyItems: questionList }));
+    console.log(formUploadInfo);
+    open({
+      title: '폼 생성',
+      desc: '폼을 생성하시겠습니까?',
+      option: {
+        type: 'CONFIRM',
+        confirmEvent: () => {
+          createForm(formUploadInfo);
+        },
+      },
+    });
+  };
+
+  useEffect(() => {
+    useTopBarStore.setState({
+      isBackButtonVisible: true,
+      isNotificationButtonVisible: false,
+    });
+  }, []);
+
+  function reducer(
+    state: Question[],
+    action: FormUploadActionType,
+  ): Question[] {
     switch (action.type) {
       case 'handleInputChange':
         return state.map((questionValue, qIndex) =>
@@ -100,189 +128,45 @@ export const FormUploadPage = () => {
     }
   }
 
-  const periodState = useState<Period>({
-    start: new Date(),
-    end: new Date(),
-  });
-
-  const formUploadInfoState = useState<FormUpload>({
-    title: '',
-    description: '',
-    postId: 0,
-    startTime: '',
-    endTime: '',
-    surveyItems: questionList,
-  });
-
-  const [formUploadInfo, setFormUploadInfo] = formUploadInfoState;
-
-  const { open } = useModal();
-  const { mutate: createForm } = useCreateForm();
-
-  const uploadForm = () => {
-    setFormUploadInfo((prev) => ({ ...prev, surveyItems: questionList }));
-    console.log(formUploadInfo);
-    open({
-      title: '폼 생성',
-      desc: '폼을 생성하시겠습니까?',
-      option: {
-        type: 'CONFIRM',
-        confirmEvent: () => {
-          createForm(formUploadInfo);
-        },
-      },
-    });
-  };
-
-  useEffect(() => {
-    useTopBarStore.setState({
-      isBackButtonVisible: true,
-      isNotificationButtonVisible: false,
-    });
-  }, []);
-
-  type QuestionTypeButton = { type: QuestionType; icon: ReactNode };
-  const questionTypeList: QuestionTypeButton[] = [
-    { type: 'SUBJECTIVE', icon: <BsTextareaT /> },
-    { type: 'MULTIPLE', icon: <RiListRadio /> },
-  ];
   return (
     <div className="flex flex-col gap-5">
-      <Header
+      <FormHeader
         periodState={periodState}
         formUploadInfoState={formUploadInfoState}
       />
-      {questionList.map(({ tag: questionType, options }, questionIndex) => (
+      {questionList.map(({ tag: questionType }, questionIndex) => (
         <Box className="flex flex-col gap-3" key={`question-${questionIndex}`}>
-          <div className="flex justify-between text-zinc-400">
-            <div className="flex gap-3">
-              {questionTypeList.map(({ type, icon }) => (
-                <QuestionTypeButton
-                  index={questionIndex}
-                  icon={icon}
-                  value={type}
-                  checked={questionList[questionIndex].tag === type}
-                  onChange={(e: QuestionTypeButtonChangeEvent) => {
-                    dispatch({
-                      type: 'changeQuestionType',
-                      questionIndex,
-                      questionType: e.target.value,
-                    });
-                  }}
-                />
-              ))}
-            </div>
-            <button>
-              <RxCross2
-                onClick={() => {
-                  dispatch({
-                    type: 'deleteQuestion',
-                    questionIndex,
-                  });
-                }}
-              />
-            </button>
-          </div>
-          <input
-            type="text"
+          <QuestionHeader
+            questionIndex={questionIndex}
+            questionList={questionList}
+            dispatch={dispatch}
+          />
+          <QuestionInput
             name="title"
+            questionIndex={questionIndex}
+            questionList={questionList}
+            dispatch={dispatch}
             placeholder="질문을 입력해주세요"
-            value={questionList[questionIndex].title}
-            onChange={(e) => {
-              dispatch({
-                type: 'handleInputChange',
-                questionIndex,
-                inputName: e.target.name,
-                inputValue: e.target.value,
-              });
-            }}
           />
-          <input
-            type="text"
+          <QuestionInput
             name="description"
-            className="ml-[-35px] scale-[0.8] text-[16px] text-zinc-500"
+            questionIndex={questionIndex}
+            questionList={questionList}
+            dispatch={dispatch}
             placeholder="설명을 입력해주세요"
-            value={questionList[questionIndex].description}
-            onChange={(e) => {
-              dispatch({
-                type: 'handleInputChange',
-                questionIndex,
-                inputName: e.target.name,
-                inputValue: e.target.value,
-              });
-            }}
           />
-          {questionType === 'SUBJECTIVE' ? (
-            <input
-              type="text"
-              disabled
-              placeholder="답변"
-              className="rounded-lg border-[1px] border-solid bg-zinc-50 p-2 text-sm"
+
+          {questionType === 'SUBJECTIVE' && <SubjectiveInput />}
+          {questionType === 'MULTIPLE' && (
+            <OptionInputList
+              dispatch={dispatch}
+              questionList={questionList}
+              questionIndex={questionIndex}
             />
-          ) : (
-            <div className="flex w-full flex-col gap-2">
-              {options.map((option, optionIndex) => (
-                <label className="flex w-full items-center justify-between text-[16px]">
-                  <div>
-                    <input type="radio" disabled />
-                    <input
-                      type="text"
-                      value={option}
-                      onChange={(e) => {
-                        dispatch({
-                          type: 'handleOptionChange',
-                          questionIndex,
-                          optionIndex,
-                          optionValue: e.target.value,
-                        });
-                      }}
-                      placeholder="옵션을 입력해주세요"
-                      className="ml-[-10px] scale-[0.8]"
-                    />
-                  </div>
-                  <button className="text-zinc-400">
-                    <PiTrashLight
-                      onClick={() => {
-                        dispatch({
-                          type: 'deleteOption',
-                          questionIndex,
-                          optionIndex,
-                        });
-                      }}
-                    />
-                  </button>
-                </label>
-              ))}
-              <button
-                onClick={() => {
-                  dispatch({
-                    type: 'createOption',
-                    questionIndex: questionIndex,
-                  });
-                }}
-                className="flex cursor-pointer items-center text-[16px]"
-              >
-                <input type="radio" disabled />
-                <p className="ml-[-1px] scale-[0.8] text-zinc-400">
-                  옵션 추가하기
-                </p>
-              </button>
-            </div>
           )}
         </Box>
       ))}
-      <div className="flex w-full justify-end">
-        <button
-          onClick={() => {
-            dispatch({
-              type: 'createQuestion',
-            });
-          }}
-          className="btn fixed bottom-[70px] flex h-[60px] w-[60px] items-center justify-center rounded-full bg-white p-3 text-[30rem] font-bold text-primary shadow-[1px_2px_10px_0px_rgba(0,0,0,0.1)]"
-        >
-          <TbPlus size={30} />
-        </button>
-      </div>
+      <AddQuestionButton dispatch={dispatch} />
       <SubmitButton content="업로드" onClick={() => uploadForm()} />
     </div>
   );
