+++ src/pages/form/upload/components/input/SubjectiveInput.tsx
@@ -1,0 +1,8 @@
+export const SubjectiveInput = () => (
+  <input
+    type="text"
+    disabled
+    placeholder="답변"
+    className="rounded-lg border-[1px] border-solid bg-zinc-50 p-2 text-sm"
+  />
+);
